Refactor: Extract geometry logic to domain/geometry.js
- Extracted geometry functions (toSvgY, generateSvgPath, getRotatedAABB, etc.) from utils.js to domain/geometry.js.
- Updated utils.js to re-export functions for backward compatibility.
- Added docs/refactoring/ANALYSIS_AND_PLAN.md to document system architecture and refactoring plan.
